requirements.txt
flask
flask-cors
google-generativeai
python-dotenv
yfinance
pandas
requests
gunicorn
